Backend-CRM/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

pydantic==2.5.0
pydantic-settings==2.1.0

celery[redis]==5.3.4
redis==4.6.0

python-multipart==0.0.6
websockets==12.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

nest-asyncio==1.5.8
google-generativeai>=0.8.0
httpx==0.25.2

motor==3.3.2
pymongo==4.6.1

shortuuid==1.0.11
weasyprint==60.2
reportlab==4.0.4
pdfplumber==0.10.3
mammoth==1.6.0
python-docx==1.1.0
pypandoc==1.13
PyJWT==2.8.0